pyproject.toml
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.poetry]
name = "shortner"
version = "0.2.5"
description = ""
authors = ["Kirill Sumorokov <sumorokov.k@gmail.com>"]
packages = [
    { include = "shortner", from = "src" },
]

[tool.poetry.dependencies]
aiohttp-micro = {git = "https://github.com/clayman083/aiohttp-micro.git", rev = "v0.3.0"}
aiohttp-metrics = {git = "https://github.com/clayman083/aiohttp-metrics.git", rev = "v0.1.3"}
aiohttp-openapi = {git = "https://github.com/clayman083/aiohttp-openapi.git", rev = "v0.1.1"}

python = "^3.8"

[tool.poetry.dev-dependencies]
ansible = "2.9.13"
black = "^19.10b0"
coverage = "*"
faker = "*"
flake8 = "*"
flake8-bugbear = "*"
flake8-builtins-unleashed = "*"
flake8-comprehensions = "*"
flake8-import-order = "*"
flake8-print = "*"
flake8-pytest = "*"
jinja2 = "2.11.2"
mypy = "*"
pytest = "*"
pytest-aiohttp = "*"
rope = "*"
